+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/AccountFeedContentStates.kt
@@ -80,10 +80,10 @@ class AccountFeedContentStates(
     val closedPollsFeed = FeedContentState(ClosedPollsFeedFilter(account), scope, LocalCache)
 
     val picturesFeed = FeedContentState(PictureFeedFilter(account), scope, LocalCache)
-    val shortsFeed = FeedContentState(ShortsFeedFilter(account), scope, LocalCache)
-    val longsFeed = FeedContentState(LongsFeedFilter(account), scope, LocalCache)
+    val shortsFeed = FeedContentState(ShortsFeedFilter(account, LocalCache), scope, LocalCache)
+    val longsFeed = FeedContentState(LongsFeedFilter(account, LocalCache), scope, LocalCache)
 
-    val notifications = CardFeedContentState(NotificationFeedFilter(account), scope)
+    val notifications = CardFeedContentState(NotificationFeedFilter(account, LocalCache), scope)
     val notificationsOpenPolls = OpenPollsState(account, scope)
     val notificationSummary = NotificationSummaryState(account)
