+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/followPacks/feed/dal/FollowPackFeedConversationsFeedFilter.kt
@@ -20,101 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.followPacks.feed.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows.AllUserFollowsByOutboxTopNavFilter
-import com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows.AllUserFollowsByProxyTopNavFilter
-import com.vitorpamplona.amethyst.model.topNavFeeds.noteBased.muted.MutedAuthorsByOutboxTopNavFilter
-import com.vitorpamplona.amethyst.model.topNavFeeds.noteBased.muted.MutedAuthorsByProxyTopNavFilter
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.amethyst.ui.dal.FilterByListParams
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip01Core.core.Event
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
-import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
-import com.vitorpamplona.quartz.nip53LiveActivities.chat.LiveActivitiesChatMessageEvent
-import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceReplyEvent
-import com.vitorpamplona.quartz.nipA4PublicMessages.PublicMessageEvent
-
-class FollowPackFeedConversationsFeedFilter(
-    val followPackNote: AddressableNote,
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + account.settings.defaultHomeFollowList.value
-
-    override fun showHiddenKey(): Boolean =
-        account.liveHomeFollowLists.value is MutedAuthorsByOutboxTopNavFilter ||
-            account.liveHomeFollowLists.value is MutedAuthorsByProxyTopNavFilter
-
-    override fun feed(): List<Note> {
-        val filterParams = buildFilterParams(account)
-
-        return sort(
-            LocalCache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it, filterParams)
-            },
-        )
-    }
-
-    val followPackEvent = followPackNote.event as? FollowListEvent
-    val follows = followPackEvent?.followIdSet() ?: emptySet()
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    fun buildFilterParams(account: Account): FilterByListParams =
-        FilterByListParams.create(
-            followLists =
-                if (account.proxyRelayList.flow.value
-                        .isEmpty()
-                ) {
-                    AllUserFollowsByOutboxTopNavFilter(
-                        authors = follows,
-                        defaultRelays = account.defaultGlobalRelays.flow,
-                        blockedRelays = account.blockedRelayList.flow,
-                    )
-                } else {
-                    AllUserFollowsByProxyTopNavFilter(
-                        authors = follows,
-                        proxyRelays = account.proxyRelayList.flow.value,
-                    )
-                },
-            hiddenUsers = account.hiddenUsers.flow.value,
-        )
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        val filterParams = buildFilterParams(account)
-
-        return collection.filterTo(HashSet()) {
-            acceptableEvent(it, filterParams)
-        }
-    }
-
-    fun acceptableEvent(
-        event: Event?,
-        relays: List<NormalizedRelayUrl>,
-        filterParams: FilterByListParams,
-    ): Boolean =
-        (
-            event is TextNoteEvent ||
-                event is ZapPollEvent ||
-                event is ChannelMessageEvent ||
-                event is CommentEvent ||
-                event is VoiceReplyEvent ||
-                event is PublicMessageEvent ||
-                event is LiveActivitiesChatMessageEvent
-        ) &&
-            filterParams.match(event, relays)
-
-    fun acceptableEvent(
-        note: Note,
-        filterParams: FilterByListParams,
-    ): Boolean = acceptableEvent(note.event, note.relays, filterParams) && !note.isNewThread()
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias FollowPackFeedConversationsFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.FollowPackFeedConversationsFeedFilter
