+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/followPacks/feed/dal/FollowPackFeedNewThreadFeedFilter.kt
@@ -20,135 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.followPacks.feed.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.filterIntoSet
-import com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows.AllUserFollowsByOutboxTopNavFilter
-import com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows.AllUserFollowsByProxyTopNavFilter
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.amethyst.ui.dal.FilterByListParams
-import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
-import com.vitorpamplona.quartz.experimental.audio.track.AudioTrackEvent
-import com.vitorpamplona.quartz.experimental.interactiveStories.InteractiveStoryPrologueEvent
-import com.vitorpamplona.quartz.experimental.nipsOnNostr.NipTextEvent
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip18Reposts.GenericRepostEvent
-import com.vitorpamplona.quartz.nip18Reposts.RepostEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
-import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
-import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
-import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
-import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceEvent
-
-class FollowPackFeedNewThreadFeedFilter(
-    val followPackNote: AddressableNote,
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    companion object {
-        val ADDRESSABLE_KINDS =
-            listOf(
-                AudioTrackEvent.KIND,
-                InteractiveStoryPrologueEvent.KIND,
-                WikiNoteEvent.KIND,
-                NipTextEvent.KIND,
-                ClassifiedsEvent.KIND,
-                LongTextNoteEvent.KIND,
-            )
-    }
-
-    val followPackEvent = followPackNote.event as? FollowListEvent
-    val follows = followPackEvent?.followIdSet() ?: emptySet()
-
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + followPackNote.idHex
-
-    override fun showHiddenKey(): Boolean = false
-
-    fun buildFilterParams(account: Account): FilterByListParams =
-        FilterByListParams.create(
-            followLists =
-                if (account.proxyRelayList.flow.value
-                        .isEmpty()
-                ) {
-                    AllUserFollowsByOutboxTopNavFilter(
-                        authors = follows,
-                        defaultRelays = account.defaultGlobalRelays.flow,
-                        blockedRelays = account.blockedRelayList.flow,
-                    )
-                } else {
-                    AllUserFollowsByProxyTopNavFilter(
-                        authors = follows,
-                        proxyRelays = account.proxyRelayList.flow.value,
-                    )
-                },
-            hiddenUsers = account.hiddenUsers.flow.value,
-        )
-
-    override fun feed(): List<Note> {
-        val filterParams = buildFilterParams(account)
-
-        val notes =
-            LocalCache.notes.filterIntoSet { _, note ->
-                // Avoids processing addressables twice.
-                (note.event?.kind ?: 99999) < 10000 && acceptableEvent(note, filterParams)
-            }
-
-        val longFormNotes =
-            LocalCache.addressables.filterIntoSet(
-                kinds = ADDRESSABLE_KINDS,
-            ) { _, note ->
-                acceptableEvent(note, filterParams)
-            }
-
-        return sort(notes + longFormNotes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        val filterParams = buildFilterParams(account)
-
-        return collection.filterTo(HashSet()) {
-            acceptableEvent(it, filterParams)
-        }
-    }
-
-    private fun acceptableEvent(
-        it: Note,
-        filterParams: FilterByListParams,
-    ): Boolean {
-        val noteEvent = it.event
-        return (
-            noteEvent is TextNoteEvent ||
-                noteEvent is ClassifiedsEvent ||
-                noteEvent is RepostEvent ||
-                noteEvent is GenericRepostEvent ||
-                (noteEvent is LongTextNoteEvent && noteEvent.content.isNotEmpty()) ||
-                (noteEvent is WikiNoteEvent && noteEvent.content.isNotEmpty()) ||
-                noteEvent is ZapPollEvent ||
-                noteEvent is HighlightEvent ||
-                noteEvent is InteractiveStoryPrologueEvent ||
-                noteEvent is CommentEvent ||
-                noteEvent is AudioTrackEvent ||
-                noteEvent is VoiceEvent ||
-                noteEvent is AudioHeaderEvent
-        ) &&
-            filterParams.match(noteEvent, it.relays) &&
-            it.isNewThread()
-    }
-
-    override fun sort(items: Set<Note>): List<Note> =
-        items
-            .distinctBy {
-                if (it.event is RepostEvent || it.event is GenericRepostEvent) {
-                    it.replyTo?.lastOrNull()?.idHex ?: it.idHex // only the most recent repost per feed.
-                } else {
-                    it.idHex
-                }
-            }.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias FollowPackFeedNewThreadFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.FollowPackFeedNewThreadFeedFilter
