+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/followPacks/feed/dal/FollowPackFeedNewThreadFeedViewModel.kt
@@ -24,12 +24,13 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.vitorpamplona.amethyst.model.Account
 import com.vitorpamplona.amethyst.model.AddressableNote
+import com.vitorpamplona.amethyst.model.LocalCache
 import com.vitorpamplona.amethyst.ui.screen.AndroidFeedViewModel
 
 class FollowPackFeedNewThreadFeedViewModel(
     val note: AddressableNote,
     val account: Account,
-) : AndroidFeedViewModel(FollowPackFeedNewThreadFeedFilter(note, account)) {
+) : AndroidFeedViewModel(FollowPackFeedNewThreadFeedFilter(note, account, LocalCache)) {
     class Factory(
         val note: AddressableNote,
         val account: Account,
