+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/longs/dal/LongsFeedFilter.kt
@@ -20,99 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.longs.dal
 
-import com.vitorpamplona.amethyst.commons.richtext.RichTextParser
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.TopFilter
-import com.vitorpamplona.amethyst.model.filterIntoSet
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.amethyst.ui.dal.FilterByListParams
-import com.vitorpamplona.amethyst.ui.screen.loggedIn.video.dal.SupportedContent
-import com.vitorpamplona.amethyst.ui.screen.loggedIn.video.datasource.SUPPORTED_VIDEO_FEED_MIME_TYPES_SET
-import com.vitorpamplona.quartz.nip01Core.core.AddressableEvent
-import com.vitorpamplona.quartz.nip71Video.VideoHorizontalEvent
-import com.vitorpamplona.quartz.nip71Video.VideoMeta
-import com.vitorpamplona.quartz.nip71Video.VideoNormalEvent
-import kotlin.collections.plus
-
-class LongsFeedFilter(
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + followList().code
-
-    override fun limit() = 200
-
-    fun followList(): TopFilter = account.settings.defaultLongsFollowList.value
-
-    fun TopFilter.isMuteList() = this is TopFilter.MuteList
-
-    fun TopFilter.isBlockList() = this is TopFilter.PeopleList && this.address == account.blockPeopleList.getBlockListAddress()
-
-    fun TopFilter.wantsToSeeNegativeStuff() = isMuteList() || isBlockList()
-
-    override fun showHiddenKey(): Boolean = followList().wantsToSeeNegativeStuff()
-
-    override fun feed(): List<Note> {
-        val params = buildFilterParams(account)
-        val notes =
-            LocalCache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it, params)
-            } +
-                LocalCache.addressables.filterIntoSet(VideoHorizontalEvent.KIND) { _, it ->
-                    acceptableEvent(it, params)
-                }
-        return sort(notes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    fun buildFilterParams(account: Account): FilterByListParams =
-        FilterByListParams.create(
-            account.liveLongsFollowLists.value,
-            account.hiddenUsers.flow.value,
-        )
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        val params = buildFilterParams(account)
-
-        return collection.filterTo(HashSet()) {
-            acceptableEvent(it, params)
-        }
-    }
-
-    val videoFeedSupport =
-        SupportedContent(
-            blockedUrls = listOf("youtu.be", "youtube.com"),
-            mimeTypes = SUPPORTED_VIDEO_FEED_MIME_TYPES_SET,
-            supportedFileExtensions = (RichTextParser.videoExtensions + RichTextParser.imageExtensions).toSet(),
-        )
-
-    fun acceptableVideoiMetas(iMetas: List<VideoMeta>): Boolean = iMetas.any { videoFeedSupport.acceptableUrl(it.url, it.mimeType) }
-
-    fun acceptanceEvent(noteEvent: VideoHorizontalEvent) = acceptableVideoiMetas(noteEvent.imetaTags())
-
-    fun acceptanceEvent(noteEvent: VideoNormalEvent) = acceptableVideoiMetas(noteEvent.imetaTags())
-
-    fun acceptableEvent(
-        note: Note,
-        params: FilterByListParams,
-    ): Boolean {
-        val noteEvent = note.event
-
-        if (noteEvent is AddressableEvent && note !is AddressableNote) {
-            return false
-        }
-
-        return (
-            (noteEvent is VideoHorizontalEvent && acceptanceEvent(noteEvent)) ||
-                (noteEvent is VideoNormalEvent && acceptanceEvent(noteEvent))
-        ) &&
-            params.match(noteEvent, note.relays) &&
-            (params.isHiddenList || account.isAcceptable(note))
-    }
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias LongsFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.LongsFeedFilter
