+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/gallery/dal/UserProfileGalleryFeedFilter.kt
@@ -20,81 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.gallery.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.model.filter
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.amethyst.ui.dal.FilterByListParams
-import com.vitorpamplona.quartz.experimental.profileGallery.ProfileGalleryEntryEvent
-import com.vitorpamplona.quartz.nip68Picture.PictureEvent
-import com.vitorpamplona.quartz.nip71Video.RegularVideoEvent
-import com.vitorpamplona.quartz.nip71Video.ReplaceableVideoEvent
-import com.vitorpamplona.quartz.nip71Video.VideoVerticalEvent
-
-class UserProfileGalleryFeedFilter(
-    val user: User,
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + "ProfileGallery"
-
-    override fun feed(): List<Note> {
-        val params = buildFilterParams(account)
-
-        val notes =
-            LocalCache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it, params, user)
-            }
-
-        val addressableNotes =
-            LocalCache.addressables
-                .filter(
-                    listOf(VideoVerticalEvent.KIND, VideoVerticalEvent.KIND),
-                    user.pubkeyHex,
-                ) { _, it ->
-                    acceptableEvent(it, params, user)
-                }
-
-        return sort(addressableNotes + notes).toList()
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        val params = buildFilterParams(account)
-
-        return collection.filterTo(HashSet()) { acceptableEvent(it, params, user) }
-    }
-
-    fun acceptableEvent(
-        it: Note,
-        params: FilterByListParams,
-        user: User,
-    ): Boolean {
-        val noteEvent = it.event
-        return (
-            (
-                it.event?.pubKey == user.pubkeyHex &&
-                    (
-                        noteEvent is PictureEvent ||
-                            noteEvent is RegularVideoEvent ||
-                            (noteEvent is ReplaceableVideoEvent && it is AddressableNote) ||
-                            (noteEvent is ProfileGalleryEntryEvent && noteEvent.hasUrl() && noteEvent.hasFromEvent())
-                    )
-            ) // && noteEvent.isOneOf(SUPPORTED_VIDEO_FEED_MIME_TYPES_SET))
-        ) &&
-            params.match(noteEvent, it.relays) &&
-            account.isAcceptable(it)
-    }
-
-    fun buildFilterParams(account: Account): FilterByListParams =
-        FilterByListParams.create(
-            followLists = account.liveStoriesFollowLists.value,
-            hiddenUsers = account.hiddenUsers.flow.value,
-        )
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias UserProfileGalleryFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.UserProfileGalleryFeedFilter
