+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/FollowPackFeedNewThreadFeedFilter.kt
@@ -1,0 +1,135 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.ui.feeds
+
+import com.vitorpamplona.amethyst.commons.model.AddressableNote
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.Note
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
+import com.vitorpamplona.quartz.experimental.audio.track.AudioTrackEvent
+import com.vitorpamplona.quartz.experimental.interactiveStories.InteractiveStoryPrologueEvent
+import com.vitorpamplona.quartz.experimental.nipsOnNostr.NipTextEvent
+import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
+import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
+import com.vitorpamplona.quartz.nip18Reposts.GenericRepostEvent
+import com.vitorpamplona.quartz.nip18Reposts.RepostEvent
+import com.vitorpamplona.quartz.nip22Comments.CommentEvent
+import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
+import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
+import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
+import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
+import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
+import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceEvent
+
+class FollowPackFeedNewThreadFeedFilter(
+    val followPackNote: AddressableNote,
+    val account: IAccount,
+    val cacheProvider: ICacheProvider,
+) : AdditiveFeedFilter<Note>() {
+    companion object {
+        val ADDRESSABLE_KINDS =
+            listOf(
+                AudioTrackEvent.KIND,
+                InteractiveStoryPrologueEvent.KIND,
+                WikiNoteEvent.KIND,
+                NipTextEvent.KIND,
+                ClassifiedsEvent.KIND,
+                LongTextNoteEvent.KIND,
+            )
+    }
+
+    val followPackEvent = followPackNote.event as? FollowListEvent
+    val follows = followPackEvent?.followIdSet() ?: emptySet()
+
+    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + followPackNote.idHex
+
+    override fun showHiddenKey(): Boolean = false
+
+    fun buildFilterParams(account: IAccount): FilterByListParams =
+        FilterByListParams.create(
+            followLists = account.createFollowListFilter(follows),
+            hiddenUsers = account.liveHiddenUsers,
+        )
+
+    override fun feed(): List<Note> {
+        val filterParams = buildFilterParams(account)
+
+        val notes =
+            cacheProvider.filterNotesIntoSet { _, note ->
+                // Avoids processing addressables twice.
+                (note.event?.kind ?: 99999) < 10000 && acceptableEvent(note, filterParams)
+            }
+
+        val longFormNotes =
+            cacheProvider.filterAddressablesIntoSet(
+                kinds = ADDRESSABLE_KINDS,
+            ) { _, note ->
+                acceptableEvent(note, filterParams)
+            }
+
+        return sort(notes + longFormNotes)
+    }
+
+    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
+
+    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
+        val filterParams = buildFilterParams(account)
+
+        return collection.filterTo(HashSet()) {
+            acceptableEvent(it, filterParams)
+        }
+    }
+
+    private fun acceptableEvent(
+        it: Note,
+        filterParams: FilterByListParams,
+    ): Boolean {
+        val noteEvent = it.event
+        return (
+            noteEvent is TextNoteEvent ||
+                noteEvent is ClassifiedsEvent ||
+                noteEvent is RepostEvent ||
+                noteEvent is GenericRepostEvent ||
+                (noteEvent is LongTextNoteEvent && noteEvent.content.isNotEmpty()) ||
+                (noteEvent is WikiNoteEvent && noteEvent.content.isNotEmpty()) ||
+                noteEvent is ZapPollEvent ||
+                noteEvent is HighlightEvent ||
+                noteEvent is InteractiveStoryPrologueEvent ||
+                noteEvent is CommentEvent ||
+                noteEvent is AudioTrackEvent ||
+                noteEvent is VoiceEvent ||
+                noteEvent is AudioHeaderEvent
+        ) &&
+            filterParams.match(noteEvent, it.relays) &&
+            it.isNewThread()
+    }
+
+    override fun sort(items: Set<Note>): List<Note> =
+        items
+            .distinctBy {
+                if (it.event is RepostEvent || it.event is GenericRepostEvent) {
+                    it.replyTo?.lastOrNull()?.idHex ?: it.idHex
+                } else {
+                    it.idHex
+                }
+            }.sortedWith(DefaultFeedOrder)
+}
